[calcabrina] Add experimental widescreen support.
This commit also revises the base window size from the original default
of 256, 224 to adjusted values (depending on if widescreen or not). This
has a few minor consequences:

First, the output will no longer exactly match the original SNES output.
This was never going to be the case, however.

Second, the horizontal component of the base resolution is no longer an
integer. This will potentially require some sort of mitigation at some
point, but with the existing code to draw outdoor maps, it doesn't
affect anything.

As a result, however, the base number of lines is now 240, so the output
can be properly scaled to common HD resolutions with an integer.

Signed-off-by: Jason Lynch <jason@calindora.com>